+++ NEWS.md
@@ -18,6 +18,9 @@
   bootstrapped confidence intervals. The function gains following new arguments:
   `ci`, `ci_method` and `iterations`.
 
+* New function `r2_ferrari()` to compute Ferrari & Cribari-Neto's R2 for 
+  generalized linear models, in particular beta-regression.
+
 # performance 0.12.0
 
 ## Breaking